requirements.txt
requests
pyyaml
psutil
pywinauto
Pillow
